checkpoint

+++ experiments/data.py
@@ -28,28 +28,28 @@ class ExperimentData:
     @staticmethod
     def load(path: str):
         experiment_data = torch.load(path)
-        experiment_data.full.x.to(torch.double)
-        experiment_data.full.y.to(torch.double)
+        experiment_data.full.x
+        experiment_data.full.y
         if experiment_data.full.y_untransformed is not None:
-            experiment_data.full.y_untransformed.to(torch.double)
+            experiment_data.full.y_untransformed
         experiment_data.full.name = "full"
         if experiment_data.train is not None:
-            experiment_data.train.x.to(torch.double)
-            experiment_data.train.y.to(torch.double)
+            experiment_data.train.x
+            experiment_data.train.y
             if experiment_data.train.y_untransformed is not None:
-                experiment_data.train.y_untransformed.to(torch.double)
+                experiment_data.train.y_untransformed
             experiment_data.train.name = "train"
         if experiment_data.test is not None:
-            experiment_data.test.x.to(torch.double)
-            experiment_data.test.y.to(torch.double)
+            experiment_data.test.x
+            experiment_data.test.y
             if experiment_data.test.y_untransformed is not None:
-                experiment_data.test.y_untransformed.to(torch.double)
+                experiment_data.test.y_untransformed
             experiment_data.test.name = "test"
         if experiment_data.validation is not None:
-            experiment_data.validation.x.to(torch.double)
-            experiment_data.validation.y.to(torch.double)
+            experiment_data.validation.x
+            experiment_data.validation.y
             if experiment_data.validation.y_untransformed is not None:
-                experiment_data.validation.y_untransformed.to(torch.double)
+                experiment_data.validation.y_untransformed
             experiment_data.validation.name = "validation"
         experiment_data.y_mean = float(experiment_data.y_mean.detach().item())
         experiment_data.y_std = float(experiment_data.y_std.detach().item())
